Improved map display

+++ patients-transport-center/src/main/java/pl/group2/optimizer/impl/items/patients/Patients.java
@@ -93,7 +93,7 @@ public class Patients implements Items {
     }
 
     @Override
-    public void draw(Graphics g, int scalaX, int scalaY, int minX, int minY) {
+    public void draw(Graphics g, double scalaX, double scalaY, int minX, int minY) {
 
     }
 }
